[CodeGen] emit atomic statements
Calls to `transaction_start()` and `transaction_end()` are inserted to the start and end of each transaction respectively. `transaction_start()` is emitted like a simple call expression directly as soon as the program enters a transaction; `transaction_end()`’s emission is deferred, piggybacking on the transaction scope’s clean up operation, so that it’s always called when program exits a transaction scope, and that stack objects declared in the scope doesn’t out-live the transaction.